Rename behaviors package to itemBehaviors, Behavior to ItemBehavior
