+++ core/src/main/scala/io/github/scalaquest/core/model/behaviorBased/common/package.scala
@@ -2,8 +2,8 @@ package io.github.scalaquest.core.model.behaviorBased
 
 /**
  * Contains a set of common [[BehaviorBasedModel.BehaviorBasedItem]],
- * [[BehaviorBasedModel.Behavior]], [[BehaviorBasedModel.GroundBehavior]] of common use, accessible
- * with a mixim mechanism from the [[BehaviorBasedModel]].
+ * [[BehaviorBasedModel.ItemBehavior]], [[BehaviorBasedModel.GroundBehavior]] of common use,
+ * accessible with a mixim mechanism from the [[BehaviorBasedModel]].
  *
  * Usage example:
  * {{{
