+++ core/src/test/scala/io/github/scalaquest/core/model/behaviorBased/common/itemBehaviors/impl/SimpleEatableTest.scala
@@ -1,4 +1,4 @@
-package io.github.scalaquest.core.model.behaviorBased.common.behaviors.impl
+package io.github.scalaquest.core.model.behaviorBased.common.itemBehaviors.impl
 
 import io.github.scalaquest.core.TestsUtils.{simpleState, startRoom}
 import io.github.scalaquest.core.model.Action.Common.Eat
